// frontend/shared/types/api.ts
export interface Company {
    id: number;
    name: string;
    street_address: string;
    city: string;
    state: string;
    zip_code: string;
    country: string;
    phone_number: string;
}
// ...
export interface Campaign {
    id: number;
    name: string;
    description?: string;
    campaign_type: string;
    publishing_type?: string[];
    landing_page_url?: string;
    tracking_pixel?: string;
    ip_precedence?: string;
    location_precedence?: string;
    locale_precedence?: string;
    browser_precedence?: string;
    time_precedence?: string;
    ip_tracking?: string[];
    location_tracking?: string[];
    locale_tracking?: string[];
    browser_tracking?: string[];
    time_tracking?: string[];
    company?: number | Company;
}